requirements.txt
PyQt5==5.15.9
opencv-python==4.8.0.76
numpy==1.24.3
av==10.0.0
pyvirtualcam==0.12.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-qt==4.2.0
pytest-mock==3.12.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
scikit-learn==1.4.0
scikit-image==0.22.0

# Captioner dependencies
pyaudio==0.2.11
pillow==10.0.0
whisper==1.1.10
speech-recognition==3.10.0

